Fix router implementation
