+++ src/components/PokemonDetail.vue
@@ -1,8 +1,6 @@
 <template>
     <div>
         <h1>Pokemon Details</h1>
-        //<img class="sprite" :src="pokemon.sprites.front_default">
-        //descripcion detallada
         <img src="https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/4.png">
     </div>
 </template>
